move full field tiff generation utils to public place
this will make them available for use in mesoscope_cli testing
