+++ src/ophys_etl/test_utils/__init__.py
@@ -1,0 +1,2 @@
+# this directory contains utilities we use for creating realistic
+# data for unit tests
