+++ src/ophys_etl/test_utils/full_field_tiff_utils.py
